requirements.txt
boto3==1.15.18
botocore==1.18.18
Django==2.2.3
django-cors-headers==2.2.0
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.4.0
django-rest-swagger==2.2.0
django-sendsms==0.4
django-ses==1.0.3
django-shortuuidfield==0.1.3
djangorestframework==3.11.1
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
environ==1.0
flower==0.9.5
future==0.18.2
geojson==2.4.1
geopy==2.2.0
haversine==2.0.0
numpy==1.16.2
polyline==1.4.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
requests==2.21.0
reverse-geocoder==1.5.1
Shapely==1.6.4.post2
six==1.16.0
urllib3==1.22
